crud

+++ CrudUsuario/CrudUsuario.Domain/Interfaces/Repositories/IRepositoryBase.cs
@@ -7,7 +7,7 @@ namespace CrudUsuario.Domain.Interfaces.Repositories
     public interface IRepositoryBase<TEntity> where TEntity : class
     {
         TEntity Add(TEntity obj);
-        TEntity GetById(int id);
+        TEntity GetById(long id);
         IEnumerable<TEntity> GetAll();
         TEntity Update(TEntity obj);
         void Remove(TEntity obj);
