+++ CrudUsuario/CrudUsuario.Domain/Interfaces/Services/IServiceBase.cs
@@ -7,7 +7,7 @@ namespace CrudUsuario.Domain.Interfaces.Services
     public interface IServiceBase<TEntity> where TEntity : class
     {
         void Add(TEntity obj);
-        TEntity GetById(int id);
+        TEntity GetById(long id);
         IEnumerable<TEntity> GetAll();
         void Update(TEntity obj);
         void Remove(TEntity obj);
